Remove DM_FRONTEND_DOMAIN default value
There's no frontend proxy on development, so there's no value to
use for the frontend domain. Using one of the apps is misleading,
since there's no proxying in any of the dev app servers either.

It's possible to use the individual development app URLs to run
tests for one app only.