Add a support/patches.rb file with fix for current_url double escape
Poltergeist capybara driver returns `current_url` that was encoded
twice. E.g. '%' in query string becomes '%2525' instead of '%25'.

(https://github.com/teampoltergeist/poltergeist/issues/418)

This is fixed in the master branch, but not in the current 1.6 release.

The patch replaces `current_url` method with the one that returns a
decoded value by changing all instances of '%25' to '%'. This should
be removed once poltergeist is updated to the release that includes
the master fix.