Add a step to retrieve a random service data from the API
Selects a random page from the API services list and returns data
for one of the services.

Service data is stored in `@service`. Since this creates a variable
dependency between this API step and any step that relies on the
`@service` being set I suggest we use `that service` in any steps
that use `@service`, e.g.:

    Then I can search for that service
    When I enter that service.id in the 'q' field
    Then I am on a page with that service in search results

This way `that service` always refers to full service data stored in
`@service`. The service doesn't need to be random, so any steps that
create or request from API full or partial service data should store
it in `@service` as long as the key names match the ones in API
response.