Change smoke tests to use random service from the API
Instead of using a preset service that doesn't exist in production
use a randomly selected service to test buyer-frontend search and
filtering.

Since some of the steps were relying on the given service data for
some of the checks (e.g. service supplier and lot) changing them to
work with the randomly selected service required splitting some of the
steps.

Individual steps that require few fields (e.g. "Selected lot") have a
generic version that accepts the data directly and "that service"
version that takes the relevant value from `@service` data and uses it
to call the generic step.