Add create draft SCS service flow
Creates a new draft SCS service and completes the pages.

Several steps are marked as @wip because they are how the application should work rather than how it is right now.

These WIP tags can be removed once the pricing page is in and document upload pages are moved to the end of the page flow.