Remove use of eval and improve environment variables
This commit does two things but they're a bit tricky to pick apart.

1. Remove use of `eval`

`eval` is considered harmful because it allows arbitrary execution of
code from variables that would not necesarrily be expected to be running
code. Most of the uses were just calling the functions they represent
however for the dynamic domain and access token uses two new functions
had to be introduced to resolve the domain and access tokens for the
specific apps.

2. Rename environment variables to closer match apps

Rename the environment variables to closer match the field names used in
the frontend applications. This makes it clearer what the environment
variables represent when adding them to the Jenkins configuration. Also,
making the function names match the environment variables they wrap
makes it easier to follow where these variables are being used.