Remove query string check
This query string check is failing for searches including service names with brackets.

The CGI encoder encodes brackets as `%28` and `%29` but our app doesn't, so matches fail.
Using a URI encoder is worse because as well as encoding the brackets it encodes spaces as `%20` instead of `+`, which our apps do.

Doing ad-hoc replacement of `%28` and `%29` in CGI-escaped strings feels too hacky, and who knows what other characters might cause the same problem in future.

The scenario description is just that the submitted query appears in the search summary on the page - it doesn't mention anything about the query parameter, so I think removing that check from this test is OK.  It still does what the scenario demands.

The alternative would be to change the way the buyer app and search API encode and decode query strings, but as they are all working nicely as they are I don't see the point in changing them just so we can run this test.