Fix finished service check
The previous message 'The service is complete' which the functional
tests were looking for never made it into the app.
Instead, when draft services are complete, a submit button labled 'Mark
as complete' shows up on the page.
Reworked the tests to test for (the absense or presense of) the 'Mark as
complete' button.