Refactor some existing step definitions
This refactors two step definitions to use less complex XPath
selectors and introduce some new functionality.

When I click the ? link for the supplier ? is functionally the same but
changes large XPath into ruby code. The only addition it makes is to set
the supplier name into the session as it is used later.

I click the ? link for ? has changed to be closer to what the step
definition is saying. Previously is built an HREF and tried to find it.
This was brittle and did not transfer well to other uses. This change
first finds the correct summary section with the summary heading and
then finds the action link that follows it, checks it is correct and
then clicks it.