Refactor admin supplier listing by prefix step
This step definition was parameterized but it could only be called with
with specific parameters. This change makes the step explicit about that
and refactors it contents.

In the changes to it's contents it makes broadly the same checks but
with less complex XPath. It also introduces the idea of different links
being visible to different user types.