Match URLs ignoring parameter order
For our applications the order of URL parameters does not matter. Also,
because the URLs are generated from dictionaries the order of URL
parameters is non-deterministic. This change introduces a couple of new
RSpec matchers that compare two URLs based on their component parts,
ignoring the order of query or fragment parameters.