Fix buyer account creation scenario
The previous buyer account creation scenario started by navigating
directly to the create page because there was no other way to get
to that page at the time that the test was written.

Now that we have a viable user flow, we can update the tests to
reflect what a user might actually do: ie, create an account as
part of creating a brief.