Replace scripts/ with Makefile and config/ environment scripts
Removes create_users and test_dependencies scripts since their steps
should be performed by the test scenarios themselves. App status
check are always run first and they check that apps are available.
Same should be done for the framework status checks for the steps
that require specific ones.

Replacing run_tests script with a Makefile since that's what we're
using in other apps. Makefile makes it a bit harder to run a custom
set of tags, but it makes the common thing easy, which is the main
problem we had in the past.