Store failed cucumber scenarios and add a make step to rerun them
Adding an additional output to the default cucumber profile allows
us to store the failed scenario names and to rerun them using
`make rerun` step.

Note: rerun will not overwrite the failed scenarios file,
since setting the rerun output erases the file before scenarios
can be read. We could write to a different file and then use make
to move it in place, but that makes it harder to use the profile
directly.