Escape HTML when printing messages with user input
Cucumber HTML formatter doesn't escape HTML tags in printed messages,
so if there are any (and there are since we've created quite a few
when testing the apps for XSS) in the random brief/service titles they get
injected into the report page.

I've made a PR to the cucumber-ruby, but for now I'm adding the manual
escaping. This affects the CLI output, since it will go through the same
escaping as well.