Add feature for responding to briefs
The three scenarios are the happy path for responding to the three
different lots on DOS.

The scenarios check that all fields can be filled in, all radio buttons
can be chosen, the pages are correct and in the correct order and that
the correct information is replayed from briefs and service data when
required.

It also checks the correct information is shown on the summary pages at
the end of the flows.