Allow radio buttons to be targeted by question
For nice to have requirements you can't target a radio button simply by
it being 'yes' or 'no' as there may be multiple on the page.

This extends the step for choosing radio buttons to optionally accept a
question name and target the checkbox in the same fieldset as it.