Make brief responses agnostic to framework
At the moment we hardcode 'digital-outcomes-and-specialists' into
our code and fixtures. This causes issues as we now move to
DOS2. Instead of changing the hardcodings to be DOS2 (which throws
it's own set of issues because functional tests run on both
preview and staging environments which may have frameworks in
different states), we make our tests take the first DOS framework
which is live. This way we don't need to worry about framework
states across our environments.