Replace regularly failing opportunity filtering step
The `I see that brief in one of the pages that follow from clicking`
step has been failing often. This is because it loops through
many pages after another until it finds the result and pages may
randomly break for example during a deployment. This generally
doesn't cause the actual jenkins job to fail because it reruns
automatically but does make our tests a bit flakey. Therefore I
have removed tests that use this step and added assertions that all
the briefs returned on the first page of filtering match the correct
status and/or lot.

I decided to check all possible options rather than pick a random
status or lot. This makes the tests slightly more rigerous in case
the environment only contains briefs of a certain lot or status
on the first page (e.g. the first page is entirely open briefs).