Rejig fixtures to pass by value
The previous way of passing fixture service data was to pass
a ruby hash by reference to a variable inside the test code.
Unfortunately, this meant that modifying values of the hash
(eg, deleting the `developerLocations` key and adding a
`designerLocations` one) would modify the 'fixture' itself,
so modifications would persist between scenarios.

Now we call a function that returns a new hash every time.