General checkbox checking
Group all of the checkbox code in a common function so that we
don't have to pass in the options every time in each of the
steps.

Also make sure that labels are visible everytime before checking
a particular checkbox.  We never want to check a checkbox that is
completely hidden from users.