Update buyer smoke tests
Always start on the actual search page (rather than the 'G-Cloud'
lot listing page), and click the 'Filter' button.

We are removing the search bar and button on the `/g-cloud` page
for G9, which means that if we are testing the searching
behaviour, we need to be on the actual listings page.