Remove hardcoded lots from buyer smoke tests
Old behaviour presumed the existence of several lots: hardcoding
the lot name into the link to click and the breakcrumb to check.

New behaviour will look for a random lot from the latest version
of G-Cloud, and then use the lot it gets back to look for the lot
link on the `/g-cloud` page, and then look for the breadcrumb in
the `/g-cloud/search` page.

This more general behaviour will work across environments if the
lot names are different.