Remove pre G-Cloud 9 lots from page check
This appears to be a general-purpose step but in fact is only ever called with "Cloud technology and support", and would only work for that page with the pre-G-Cloud 9 lots specifically linked from it.

This removes the old lots and makes the page check overall more generic (checks only for title and breadcrumbs).