Add docker-compose file for bringing up local docker environment
The docker-compose file will start up all the apps and their
dependencies in the correct order.

The postgres image will be first, and if no container already exists it
will create a new one with a volume, and run any sql file in `./sql` to
populate the db. The intention is to have a local copy of a db dump in
here so you can easily bring up a fresh db in a known state.

The api image runs a script to wait for the postgres image to be ready
for connections. Once it is it will start the api and connect.

The frontend apps run a similar script to wait for the api to be ready
for connections before the start their apps.

The elasticsearch image will just start up and be ready. If you're bring
the environment up for the first time you'll need to index your services
using the script found in the digitalmarketplace-scripts repo.

The nginx image will route traffic as well as add basic auth headers to
satisfy the nginx instances running in front of the apps in the
containers, in the same way we do in a live environment