Add scripts for checking readiness of apps
`wait-for-postgres.py` uses psycopg2 as we don't have psql on the api
image. It checks to make sure that there is more than one framework,
which will be true once the sql import has been completed by the
postgres image startup. It DOESN'T guarantee anything about the
correctness of your DB, just that there is some data in there.