Add script to populate db from database_dump
The postgres image is built in such a way that if you have any *.sql
files in `/docker-entrypoint-initdb.d` they will be run when a container
is created. The problem is that they will also exit if an error is
encountered.

If the dump you're using was created with a `--clean` flag and `pg_dump`
then you'll get loads of errors importing it due to it not being able to
drop the tables, indexes, foreign keys etc. as they don't exist yet. And
those errors cause the import to stop.

Any *.sh scripts in the same directory will also run. We can use this to
import the dump without the annoying exit on error feature. This should
make the containers more robust to different styles of db dump.