Import db dump from a separate directory
The directory it was in before would automatically run any .sql files in
there. This meant that, because I want the `populate_db.sh` script to do
the importing, that it couldn't end in .sql.

Moving it to a directory where that doesn't happen means that it can
have the `.sql` suffix (like normal data exports) and the script can be
altered to import any file with `.sql` from that directory.