Refactor 'that supplier is on that framework' to be more verbose
Instead of one step make this 2 and add a third for returning a signed framework agreement.
Verbocity in this case makes sense because it shows the user journey.

Basically:
And that supplier is on that framework

Has become:
And 'that supplier has applied to be on that framework'
And 'we accept that suppliers application to the framework'

With an optional (but recommended):
And 'that supplier returns a signed framework agreement for the framework'

This better represents our userjourneys and shows anyone reading the tests the exact steps that are likely to happen.