Use supplier_user_id when signing framework
The api helper to sign a framework agreement was using a hard coded ID
of 1 for the uploaderUserId. This is fine if there is a user with an ID
of 1. If not it fails to sign the framework agreement and the functional
test fail at a later point.

This occurred when moving cleaned up data from prod to staging.

Passing in the supplier user id an argument to the helper solves this
issue.