Remove quotes from steps that prevents them running
There were quotes around some steps in the view_and_edit_services
scenarios. This means they weren't matching with the step definitions
and so weren't being run, just skipped. This is different to a failure
and so may have been missed.

Tests seem to pass with them being run so I'm correcting here.