Remove duplicated failing test
This test had two different versions for different environments. The
only difference was the name of a link.

Neither tests were actually being run due to the issue resolved in the
previous pull request. Now they are, it turns out the second of the
tests, the one skipped on preview and local, actually fails. The test
skipped on staging is good for all environments so we keep it and remove
the tag.