Find radio button labels using id
Due to an update in the frontend toolkit, radio button inputs are no
longer nested inside their labels. The input and the label are now both
in a div, and the label linked to the input via it's id.

We were checking that radio button's label were visible, and finding the
label by selecting the radio buttons parent. This no longer works.

Instead, we can use the radio buttons id and find all labels who's `for`
attribute is the radio button's id.