Add build instructions for local nginx to the compose file
There's no make step defined for building the nginx image locally,
but since we have a docker compose file in this repo we can us it
to run the image build.

Adding `build` to the container definition allows running
`docker-compose build` to rerun `docker build` on all defined images.
Docker will automatically tag the images with the tag from `image:`.