Update field search for frontend-toolkit v23 input/label layout
New version of frontend-toolkit form layouts put the `<input>` tag
before the `<label>` instead of nesting it inside the tag.

We need to update the XPath expressions we use to find the label
related to a given input. Instead of using the parent element, we
look for the label that has a `for` attribute set to current input's
id.