Increase processes to four for parallel tests
The default value for the number of processes used when running the
parallel tests is the number of available cpu's. This seems to be two on
Jenkins. This reduces the test run to 6 mins from 9. Not bad.

The processes will be blocked a lot waiting for I/O which means it might
be worth increasing the number of processes to 4 to try and improve on
this reduction.