Add a test for reset password flow
This adds step definitions for flows that include sending an email
to the user.

By making notification references predictable we can use Notify API
to retrieve the email by reconstructing the reference that the
application attached to it in the functional test code.

This allows us to get the email body directly from Notify.

The benefit of this approach is that we can use test Notify API
keys and avoid sending actual emails to fake email addresses.
This is also a much simpler setup since it doesn't require setting
up an email address to receive emails or adding delays to the tests
to account for email delivery.