Skip reset password tests on staging
Staging is set up to send emails and is using the production Notify
template, which will be set up to remove personalisation data.

Test Notify API key we're using doesn't seem to return live
notifications, and since the production template would remove the
reset password link from the message body the existing approach
won't work on staging. So we're only keeping the email tests on
preview.