Add a `wait` step to reset email password scenario
Preview notify tests are too quick. We create a user and request
password reset immediately. It looks like the date fernet puts inside
the token is stored with a second resolution (so it doesn’t have
milli/microseconds), but the API response does. So if the user creation
and token generation happen within the same second (which it looks like
they do quite often) the check thinks that the token was created before
the user and declares it expired.