Move wait to between user and token creation
We added a wait to ensure that there was a time difference between user
creation and token creation for password resets. When the token is
decrypted it checks that the token was created after the user.

Unfortunately I put the wait in the wrong place so we have the same
issue. Moving it hear ensures that the token is created after the user.