Add flow from homepage to mailing list functional test
The "Given there is at most one open framework" step is a shortcut to the full
logic for when the link will exist to sign up for notifications.

It will work OK while we only have G-Cloud and DOS on the marketplace, but
will need to be re-thought once there are more "meta frameworks".