Symlink `/dev/stdout` to logfile
As the containers are the same as the ones we use in production,
application logs are written to file, specifically to
`/var/log/digitalmarketplace/application.log`.

By symlinking `/dev/stdout` to the logfile, application logs which are
written to that file will actually be sent to stdout of the container,
and displayed in the terminal.