Prevent log rotation when running containers
Supervisord is set up to rotate log files. It rotates them when the file
reaches 50000000 bytes.

As the log file it's looking at is now /dev/stdout this is an issue.
Everytime a log line is written to the log file (stdout) supervisord
attempts to seek the file. Which is an illegal operation on /dev/stdout.
This causes a stack trace to be output after every log line. See here
for more info: https://github.com/Supervisor/supervisor/issues/935

By setting the logfile maxbytes to 0, we turn off log rotation which
resolves the issue.