Iterate through brief responses until response is found
We have run into problems where we call `get_brief_responses` but
end up with no responses with the required brief status. By changing
it so we can essentially pick from more than the first 100 results,
we can make it much more unlikely that we won't be able to find
a suitable response.