Update admin users email domains in config
There are now only certain domains that admin users can be created with.
They're defined in the config of the api. If you try to create or use an
admin account without one of these domains, the functional tests will
fail. If you're running tests locally, this should fix that.