Move DOS opportunities tests out of smoke-tests
The tests for the new DOS opportunities filters tests were included in
the smoke-tests as they were passive.

They more than doubled the number of requests the smoke tests were making
though, and were probably too in depth for smoke tests. The increase in
requests was causing us to breach our 0.5GB limit on logs in preview.

The detailed tests have been moved to just the functional tests.