Logical operator precedence
Ruby's `and` and `or` are not the same as Python's.

```
a = true and false
a => true

a = true && false
a => false
```

This is because of operator [precedence](https://ruby-doc.org/core-2.3.1/doc/syntax/precedence_rdoc.html)

You might see `and` and `or` used like so:

`do_something_awesome or raise RuntimeError "All the awesome is gone"`

Quite often you will also see the above written as:

`raise RuntimeError "All the awesome is gone" unless do_something_awesome`

`then` is also usually not used in conditional statements.