Change a step to look for admin email address dynamically, not statically
In the feature where we test admin users accessing supplier services
an error was introduced: one of the steps looked for a static
user email address in a message, but this email address changes
depending on the environment. So instead we are now asserting
contents of the message in two steps: one looks for "Removed by "
text while other dynamically looks for the email address.