Namespace supplier name
We were having issues with supplier names clashing when running tests in
parallel. One of the methods we use for creating suppliers does a
`get_or_create` and uses a supplier name as a prefix for the `get`. By
having very similar names we could get into a situation that a required
supplier would never be created, if a similar supplier was created
first.