Monkeypatch Capybara to catch slow finders
 ## Summary
Our tests are running a lot slower than they should. This is because we
are using the wrong Capybara finder methods in some places.

Capybara is capable of dealing with pages that have dynamic content. In
order to do this, some of its methods will look for given
elements/content, and if they cannot find it, will retry repeatedly up
to a given timeout (default 2 seconds). If we use one of these finders
to assert that text is not present on a page, then we are unnecessarily
delaying our tests because we have very few pages that rely on
changing/ajax-y content.

This monkeypatch will let us know whenever we hit a timeout on one of
our finders, which will let us work out where we're slowing our tests
down and allow us to change them to use more appropriate finders.

See
http://blog.codeship.com/faster-rails-tests?utm_source=gem_exception"
for more details.