Reduce max wait time
 ## Summary
Disable waiting for the majority of our calls to Capybara's matchers.
Our site is primarily static content after the page has loaded, so we
don't need Capybara to be willing to wait for theoretical ajax/dynamic
content to finish loading.

Where we do have dynamic content (such as live search on
services/briefs), we explicitly add a maximum wait time that Capybara
will respect.

Capybara's matchers/selectors work in a specific way to support dynamic
changes on the page:

When we ask Capybara to check for an element or some text on the page,
if the text is present, Capybara immediately notices this and returns
the result. If the text is *not* present, Capybara will wait for its
`max_default_wait_time` (seconds) configuration value, checking the page
every 0.05s to see if the content has appeared yet. After the wait time
has elapsed, if the content is still not present, Capybara raises an
'element not found' exception to report that the selector is missing.
This is all well and good if we expect the content to be there. An issue
arises when we use a 'look for this content' selector, and check that
the return value is `false`, to assert that content is *not* present. In
this case, we will always force Capybara to sleep out its
'max_default_wait_time' period. Where we want to do this, we should be
careful to use the negated matcher that Capybara often exposes (e.g.
there is a `has_content?` matcher for checking that content is present,
and a `has_no_content?` matcher for checking that content is *not*
present).

The above issue has been mitigated by setting the
'default_max_wait_time' configuration option to a low threshold (0.05
seconds). This means that, even if we use a suboptimal matcher, we won't
sleep for very long.

On top of this, we now specifically disable waiting for a number of
Capybara's matches by sending the option `wait: false` into the Capybara
matcher calls. These two effects combined will mean that it should be
reasonably unlikely for the issue to recur - however we should still be
conscientious when using Capybara's matchers in the future.

It will be important to be aware, in the future, that when we do add
further dynamic content/ajax-y stuff to the frontends, that we'll
specifically need to give a generous `wait` for the matchers (for which
we can use the new `dm_custom_wait_time` config value defined in
`features/support/env.rb`

Bump Capybara library version - required for correct passing of
`options`/`wait: false` down Capybara's matcher chain.