Remove preview/staging toggle for test
 ## Summary
https://github.com/alphagov/digitalmarketplace-brief-responses-frontend/pull/54
introduced a new form to the brief-responses-frontend that changed the
ID on one of the fields used for submitting clarification questions
about briefs, requiring a toggle on preview/staging to differentiate
between the two different versions of the code. This PR is rolling out
to staging shortly so we need to remove the toggle.