Add 'requires-credentials' tag
A number of functional tests use services that require API tokens to 
use, such as Notify and S3. While these have individual tags, it is a 
pain to add exclusions for all of them when running functional tests
without clearance and/or tokens.

This commit adds an additional `@requires-clearance` tag to all 
scenarios with `@notify`, `@mailchimp` and `@file-upload` so that they 
can be easily excluded as a group.