Fetch latest live framework
We had a step to fetch a live framework from a particular family. This
was taking the first framework returned from _all_ live frameworks for
that family. Normally that's fine as 95% of the time there is only one
live framework on a particular family. However g-cloud has an overlap.
This update means that we'll always use the latest live framework and so
we should see failures due to changes in framework state earlier.