Remove references to production users from steps
We have a number of pre-configured users for functional tests that are
used for testing functionality for the major roles; in the tests these
have been referred to as "production" users, which is misleading unless
the the tests are actually running on production.

This commit introduces a new convention of referring to these users as
existing users. Steps which refer to production users have been changed
accordingly; 'a production' -> 'an existing' etc.

While technically the user may not exist the first time the functional
tests are run, I think this is a more accurate and less alarming way of
referring to this set of specific users.