significantly refactor user creation and handling
now we actually check if the user we fetch actually has all the properties
asked for, even if the id or email address is specified. we could extend
this by, instead of aborting in cases where we have a property mismatch,
going ahead an updating the row we've found to match the request. it's
debatable how wise that is so we're leaving it for now.

we now should also be insensitive to whether user parameters are supplied
in snake_case or camelCase now.