Change search scenarios to use a new search step
Our current scenarios that test the search box assume that searches will
take place after text is entered into the search box, without the user
submitting the form. However, with our current live search javascript,
this requires the focus of the search box to be lost, which does not
always happen.

This commit changes the scenarios to explicitly submit the search form,
by adding a new pair of steps specifically related to the search box. We
don't want to rely on the form focus being lost; instead we test as if
the user were hitting enter or pressing the Search button.