853-smoke-smoulder-test-failures-on-search-results-page
We set `Capybara.default_max_wait_time` to a low threshold (0.05) to avoid waiting for a massive `default_max_wait_time` when checking that elements _do not_ exist.
Therefore if a `find` on an element raises Capybara::ElementNotFound we only wait 0.05 seconds for it to exist before returning the fact that it does not.

This is good because it means that asserting thwat an element does not exist only takes 0.05 seconds of wait.

What's bad about the previous implementation is that it also catches 2 other errors:
https://github.com/teamcapybara/capybara/blob/2.18.0/lib/capybara/node/base.rb#L125
https://github.com/teampoltergeist/poltergeist/blob/master/lib/capybara/poltergeist/driver.rb#L376
`Capybara::Poltergeist::ObsoleteNode` and `Capybara::Poltergeist::MouseEventFailed`

We are currently experiencing failures under load when only waiting 0.05 seconds on a `Capybara::Poltergeist::ObsoleteNode`.

We need to have a different wait time for Capybara::ElementNotFound and other, legitimate, page load errors.

This pull request monkeypatches Capybara::Node::Base.synchronise with a single line addition:

`seconds = e.class == Capybara::ElementNotFound ? seconds : dm_custom_wait_time`

which sets 'seconds' to our low wait time `0.05` if we're dealing with a `Capybara::ElementNotFound` error, else sets it to our standard `dm_custom_wait_time` (5s)