Add step for filtering search results by category
Our catalogue tests can be tripped up if there happens to be a search
result that contains the text of a filter or category in its name.

This commit adds a more specific step for filtering by category that
uses a CSS selector to limit where we look for the specific category
name.