Update create_supplier (properly, this time)
This will serve as a lesson for me to be more careful when changing Ruby code. The code I thought randomly generated a ten-digit number actually generated a number in the range 0-9999999999. Since we weren't validating DUNS number length, nobody noticed when a random supplier's DUNS number was, say, 3. I've replaced this with idiomatic ruby code that generates a 9-digit number. We are reasonably confident that this is kept as a string as it's passed around the system and therefore leading zeroes are a-okay.