Change service edits feature test to acknowledge all service edits
We had an issue where functional tests were failing because the number
of unacknowledge service edits on the admin page was larger than 100, so
the most recent were being pushed on to a second page and not visible.

This commit changes the service edit feature's background to make sure
that all unacknowledged service edits are acknowledged, so we can be
sure the page is clear before beginning the tests.