Allow using briefs where the user is not active
We have been running out of closed briefs with responses, partly because
we limit ourselves to briefs where the user is active; this restriction
is not necessary however. This commit adds a method
`ensure_user_can_log_in` that activates deactivated users, and we use
this in the "log in as user with brief with brief responses" step.