Change summary table steps to use Capybara Finders instead of xpath
As part of migrating to our apps to use GOV.UK Frontend buttons in forms
are going to change from input elements to button elements, which
breaks our tests because they currently search for input elements but
not button elements.

Capybara's `find_button` method takes care of this for us, as it looks
for `button` elements and `input` elements.