Update validation message step to support GOV.UK Design System
As part of migrating our frontends to use GOV.UK Design System styles
and components we are replacing instances of our validation components
with the error message component.

We want to be able to have the test working on either version, so this
commit updates the step that looks for validation messages to also look
for error messages.